Add files via upload
AllSame - generates a csv file of "students" who get every question in the test right. based on answer key of test

Dynamic Test Generator - generates a csv file of "students" with randomly chosen responses to given questions in test. dynamic because it only depends on the number of options in whatever question you give it, not the type of question.

RandomHalfBEMA - generates a csv file of "students", all of whom get half of the questions right and the answers for the others are randomly chosen. It is not dynamic as an answer key must be given. the questions each student gets right for certain is randomly chosen.

ExpectedValueCalculator - calculates the expected value for each metric we designed given random selection based on the probability of each transition occuring